Update to create a generic ClassListOfMapConverter and ClassToMapConverter
